refactor: toTally を agg/toTally.ts に切り出す
duckdbBridge から純ロジックを分離し、
aggregate() → toTally() のパイプラインを明確化。
